relocate map view and list view into posts
