+++ app/src/main/java/csis3175/w23/g11/rooftown/posts/ui/view/GridViewFragment.java
@@ -1,4 +1,4 @@
-package csis3175.w23.g11.rooftown;
+package csis3175.w23.g11.rooftown.posts.ui.view;
 
 import android.os.Bundle;
 
@@ -7,6 +7,8 @@ import androidx.fragment.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import csis3175.w23.g11.rooftown.R;
 
 public class GridViewFragment extends Fragment {
