+++ app/src/main/res/layout/fragment_list_view.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ListViewFragment">
+    tools:context=".posts.ui.view.ListViewFragment">
 
     <TextView
         android:layout_width="wrap_content"
