+++ app/src/main/res/layout/fragment_roommates.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".RoommatesFragment">
+    tools:context=".posts.ui.view.RoommatesFragment">
 
     <com.google.android.material.tabs.TabLayout
         android:layout_width="match_parent"
